addressed some comments, will remove forceLargeN next
